Cleanup of code
