+++ src/main/java/frc/robot/commands/Drive/Forward50.java
@@ -2,24 +2,21 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Drive;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Intake;
-import edu.wpi.first.wpilibj.XboxController;
-import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.subsystems.Chassis.Chassis;
 
-public class IntakeCommand extends CommandBase {
-  /** Creates a new Intake. */
-  private XboxController m_controller;
-  private Intake m_intake; 
-
-  public IntakeCommand(XboxController p, Intake k) {
-
-    m_controller = p;
-    m_intake = k;
+public class Forward50 extends CommandBase {
+  Chassis chassis;
+  double speed;
+  
+  /** Creates a new Forward50. */
+  public Forward50(Chassis chassis, double speed) {
+    this.chassis = chassis;
+    this.speed = speed;
+    addRequirements(chassis);
     // Use addRequirements() here to declare subsystem dependencies.
-    addRequirements(m_intake);
   }
 
   // Called when the command is initially scheduled.
@@ -29,20 +26,13 @@ public class IntakeCommand extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    double speed = m_controller.getRightTriggerAxis();
-    m_intake.setSpeed(speed*-.75);
-
-    if(speed > 0){
-      m_intake.DeployIntake();
-    }else{
-      m_intake.StopIntake();
-    }
+    chassis.setSpeed(speed, speed);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    m_intake.setSpeed(0);
+    chassis.setSpeed(0, 0);
   }
 
   // Returns true when the command should end.
