+++ src/main/java/frc/robot/commands/Intake/IntakeCommand.java
@@ -2,37 +2,38 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Intake;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Chassis.Chassis;
+import frc.robot.subsystems.Intake;
 
-public class Forward50 extends CommandBase {
-  Chassis chassis;
-  double speed;
-  
-  /** Creates a new Forward50. */
-  public Forward50(Chassis chassis, double speed) {
-    this.chassis = chassis;
-    this.speed = speed;
-    addRequirements(chassis);
+public class IntakeCommand extends CommandBase {
+  /** Creates a new Intake. */
+  private Intake m_intake; 
+
+  public IntakeCommand(Intake subsystem) {
+
+    m_intake = subsystem;
     // Use addRequirements() here to declare subsystem dependencies.
+    addRequirements(m_intake);
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {}
+  public void initialize() {
+    m_intake.DeployIntake();
+  }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    chassis.setSpeed(speed, speed);
+    m_intake.setSpeed(-.75);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    chassis.setSpeed(0, 0);
+    m_intake.StopIntake();
   }
 
   // Returns true when the command should end.
